Create SQL Data Cleaning.sql
This project involved cleaning and transforming data from the Uruguayan Attorney General's Office regarding accusations and convictions of individuals. The data was obtained from the official government data catalog website.

The first step was to perform data cleaning, including changing capitalization, handling null values, and standardizing gender values. The cleaned data was then ordered and filtered to generate insights.

One of the main transformations performed was the addition of a column with age ranges to better analyze age-related patterns. This was achieved by creating a new table based on the cleaned data with an additional column that categorized ages into specific ranges.

The final result was a cleaned and transformed dataset ready for further analysis and visualization. This project showcases the data cleaning and transformation skills required to turn raw data into useful and actionable insights.